Extracts interface from CombatResolver
The interface now is CombatResolver, and the implementation
is DiceRollingCombatResolver.

This change was with the intention of making the code
more testable, because we can now mock this class.
This is good though, because now the other classes
will be more decoupled from the dice rolling mechanism.
We could maybe, instead of defining it by rolling dice,
define it by just the amount of armies, or defining it by
the color of the players, or whatever.
