+++ core/src/main/combat/CombatResolver.kt
@@ -1,20 +1,9 @@
 package combat
 
 import PositiveInt
-import combat.lostArmiesCalculator.LostArmiesCalculator
 
-class CombatResolver(private val diceRoller: CombatDiceRoller) {
-
-    private val lostArmiesCalculator = LostArmiesCalculator()
-
+interface CombatResolver {
     fun armiesLostForCombat(
         attackingArmies: PositiveInt, defendingArmies: PositiveInt
-    ): Pair<Int, Int> {
-        val (attackerRolls, defenderRolls) =
-            diceRoller.forCombat(attackingArmies, defendingArmies)
-        val contestedArmies =
-            ContestedArmiesCalculator(attackingArmies, defendingArmies).getArmies()
-        return lostArmiesCalculator.armiesLostForRolls(
-            attackerRolls, defenderRolls, contestedArmies)
-    }
+    ): Pair<Int, Int>
 }
